Fix race condition when installing type support.
At a first glance, it might look like the old solution was okay because
install() is only called from static initializers. However, the §12.4.2
of The Java Language Specification (Java SE 8 Edition) only guarantees
synchronization of static initialization of the same class.
This means that the static initializers of two different classes might
run in parallel and if they both call install(), there is a race
condition that may lead to one of the two calls throwing a
RuntimeException when it cannot install the type support because it has
already been installed.